processing/requirements.txt
pika
grpcio
grpcio-tools
protobuf
psycopg2-binary
